requirements.txt
# Core data science libraries - original tested versions
pandas==2.0.3
numpy==1.26.4
matplotlib==3.9.2
seaborn==0.12.2
plotly==5.14.1
streamlit==1.25.0
scikit-learn==1.3.0
geopandas==1.0.1
contextily==1.6.2
prophet==1.1.6
cmdstanpy==1.1.0
joblib==1.3.2
openpyxl==3.1.2

# API and configuration
requests>=2.31.0
pyyaml>=6.0

# Dependencies for geopandas and contextily
shapely==2.0.6
fiona==1.9.4
pyproj==3.6.0
rasterio==1.3.8
